requirements.txt
imageio
matplotlib
numpy
pylops
scikit-image
scipy
scikit-learn
progressbar2